requirements.txt
chainlit>=0.7.700
google-generativeai>=0.3.2
python-dotenv>=0.19.0 